# doc/api/swagger.yml
swagger: "2.0"
info:
  description: "This is a test RestAPI microservice, done in Go."
  version: "1.0.0"
  title: "go-test-class-api"
  contact:
    email: "alexgrauroca91@gmail.com"
host: "localhost:5050"
basePath: "/api/v1"
...
schemes:
- "http"
paths:
  /bookings:
    get:
      tags:
      - "booking"
      summary: "Get a list of existing bookings"
      operationId: "getBookings"
      produces:
      - "application/json"
      responses:
        "200":
          description: "Bookings found"
          schema:
            type: "object"
            properties:
              bookings:
                type: array
                items:
                  $ref: "#/definitions/Booking"
        "404":
          description: "Bookings not found"
        "500":
          description: "Internal server error"
    post:
      tags:
      - "booking"
      summary: "Add a new booking"
      operationId: "addBooking"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Booking object that needs to be added"
        required: true
        schema:
          $ref: "#/definitions/CreateBooking"
      responses:
        "201":
          description: "Created booking"
          schema:
            $ref: "#/definitions/Booking"
        "400":
          description: "Bad request"
        "500":
          description: "Internal server error"
...
definitions:
  Booking:
    type: "object"
    properties:
      id:
        type: "string"
      name:
        type: "string"
      date:
        type: "string"
        format: "date-time"
...
  CreateBooking:
    type: "object"
    properties:
      name:
        type: "string"
      date:
        type: "string"
        format: "date-time"
...